+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     error management.
   - Handle unknown distribution error in server pipeline
     ([#71](https://github.com/rackslab/fatbuildr/issues/71))
+  - Handle connection error to HTTP URI with readable error message
+    ([#10](https://github.com/rackslab/fatbuildr/issues/10))
 
 ### Changed
 - docs: convert APT sources file in quickstart guide from one-line format to